refactor: merge LotSummaryTable into IncomeTable for better code organization
- Consolidate table functionality directly into IncomeTable component
- Remove unnecessary abstraction layer and improve code cohesion
- Maintain all existing functionality including sorting and filtering
- Use filtered contributions more efficiently for both summary cards and table data
- Simplify component architecture while preserving user experience
